applications/dns/requirements.txt
sanic
Sanic-WTF
gino
google-cloud-dns
